requirements.txt
websockets==12.0
numpy==2.1.1
ndi-python==5.1.1.5
pillow==10.4.0
opencv-python==4.10.0.84
imageio-ffmpeg==0.5.1
flask==3.0.3
